+++ src/utils/string.js
@@ -1,0 +1,6 @@
+export function shortenString(str, maxLen) {
+  if(str && str.length > maxLen) {
+    return str.substr(0, maxLen-1)+'...'
+  }
+  return str
+}
